requirements_py38.txt
tqdm
pandas
numpy
scipy
pillow
ipython
opencv-python
matplotlib
scikit-image
albumentations
bitsandbytes
torch==1.13.1
torchvision==0.14.1
et-xmlfile==1.1.0
openpyxl==3.1.5